+++ crates/wdl-engine/src/backend/docker.rs
@@ -231,13 +231,11 @@ impl<'a> DockerTask<'a> {
                             .expect("must have container")
                             .to_string(),
                     )
-                    .program(
-                        self.config
-                            .task
-                            .shell
-                            .as_deref()
-                            .unwrap_or(DEFAULT_TASK_SHELL),
-                    )
+                    .program(if self.config.task.shell.is_empty() {
+                        DEFAULT_TASK_SHELL
+                    } else {
+                        &self.config.task.shell
+                    })
                     .args([GUEST_COMMAND_PATH.to_string()])
                     .work_dir(GUEST_WORK_DIR)
                     .env(self.request.env.clone())
@@ -471,8 +469,15 @@ impl TaskExecutionBackend for DockerBackend {
         requirements: &HashMap<String, Value>,
         hints: &HashMap<String, Value>,
     ) -> Result<TaskExecutionConstraints> {
-        let container =
-            requirements::container(inputs, requirements, self.config.task.container.as_deref());
+        let container = requirements::container(
+            inputs,
+            requirements,
+            if self.config.task.container.is_empty() {
+                None
+            } else {
+                Some(&self.config.task.container)
+            },
+        );
         match &container {
             ContainerSource::Docker(_) => {}
             ContainerSource::Library(_) | ContainerSource::Oras(_) => {